cast noise to float, #437

+++ src/HyperspaceCloud.cpp
@@ -151,7 +151,7 @@ void HyperspaceCloud::Render(const vector3d &viewCoords, const matrix4x4d &viewT
 	// precise to the rendered frame (better than PHYSICS_HZ granularity)
 	double preciseTime = Pi::GetGameTime() + Pi::GetGameTickAlpha()*Pi::GetTimeStep();
 
-	float radius = 1000.0f + 200.0f*noise(10.0*preciseTime, 0, 0);
+	float radius = 1000.0f + 200.0f*float(noise(10.0*preciseTime, 0, 0));
 	if (m_isArrival) {
 		make_circle_thing(radius, Color(1.0,1.0,1.0,1.0), Color(0.0,0.0,1.0,0.0));
 	} else {
